move to package with its friends
